+++ examples/bootstrap/components/ColLinks.shape.php
@@ -16,7 +16,7 @@ return Compile::shape(
     div(
         h5(Slot::text('title')),
         ul(
-            Slot::each('links', Compile::shape(li(a(Slot::text('text'))->class('text-muted')->href(Slot::attr('href')))))
+            Slot::each('links', li(a(Slot::text('text'))->class('text-muted')->href(Slot::attr('href'))))
         )->class('list-unstyled text-small')
     )->class('col-6 col-md')
 );